@@ -4,8 +4,7 @@ import com.antoniomasfanclub.model.SalesRep;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
 public interface SalesRepRepository extends JpaRepository<SalesRep, Integer> {
-
-
 }
